requirements.txt

kivy
kivy_garden.mapview
requests
deutsche_bahn_api
pywin32
